[#496] Move SpanRecorderFactory from uri module to context/trace
- Move SpanRecorderFactory to lib/context/trace/span-recorder-factory.js
- Rename uri/span-recorder-factory.js to uri/uri-stats-span-recorder-factory.js
- Update all imports accordingly
