+++ index.js
@@ -13,7 +13,7 @@ const { LogBuilder } = require('./lib/utils/log/log-builder')
 const logger = require('./lib/utils/log/logger')
 const { UriStatsMonitor } = require('./lib/metric/uri/uri-stats-monitor')
 const { UriStatsConfigBuilder } = require('./lib/metric/uri/uri-stats-config-builder')
-const { UriStatsSpanRecorderFactory } = require('./lib/metric/uri/span-recorder-factory')
+const { UriStatsSpanRecorderFactory } = require('./lib/metric/uri/uri-stats-span-recorder-factory')
 const { UriStatsRepositoryBuilder } = require('./lib/metric/uri/uri-stats-repository')
 const { TraceCompletionEnricher } = require('./lib/metric/uri/trace-completion-enricher')
 const { ErrorAnalysisConfigBuilder } = require('./lib/context/trace/error-analysis-config-builder')
