+++ lib/agent-builder.js
@@ -16,7 +16,7 @@ const Scheduler = require('./utils/scheduler')
 const AgentStatsMonitor = require('./metric/agent-stats-monitor')
 const PingScheduler = require('./metric/ping-scheduler')
 const { ConfigBuilder } = require('./config-builder')
-const { SpanRecorderFactory } = require('./metric/uri/span-recorder-factory')
+const { SpanRecorderFactory } = require('./context/trace/span-recorder-factory')
 
 class Agent {
     constructor(agentInfo, dataSender, config, logger, enrichers, spanRecorderFactory) {
