+++ lib/metric/uri/uri-stats-span-recorder-factory.js
@@ -6,23 +6,8 @@
 
 'use strict'
 
-const SpanRecorder = require('../../context/trace/span-recorder')
-const DisableSpanRecorder = require('../../context/disable-span-recorder')
+const { SpanRecorderFactory } = require('../../context/trace/span-recorder-factory')
 const { SpanRecorderEnricher } = require('./span-recorder-enricher')
-
-class SpanRecorderFactory {
-    constructor(config) {
-        this.config = config
-    }
-
-    create(spanBuilder) {
-        return new SpanRecorder(spanBuilder, this.config)
-    }
-
-    createDisable(traceRoot) {
-        return new DisableSpanRecorder(traceRoot, this.config)
-    }
-}
 
 class UriStatsSpanRecorderFactory extends SpanRecorderFactory {
     constructor(config, uriStatsConfig) {
@@ -41,4 +26,4 @@ class UriStatsSpanRecorderFactory extends SpanRecorderFactory {
     }
 }
 
-module.exports = { SpanRecorderFactory, UriStatsSpanRecorderFactory }
+module.exports = { UriStatsSpanRecorderFactory }
