+++ test/support/agent-singleton-mock.js
@@ -22,7 +22,8 @@ const AgentInfo = require('../../lib/data/dto/agent-info')
 const { ConfigBuilder } = require('../../lib/config-builder')
 const { UriStatsRepositoryBuilder } = require('../../lib/metric/uri/uri-stats-repository')
 const { UriStatsConfigBuilder } = require('../../lib/metric/uri/uri-stats-config-builder')
-const { SpanRecorderFactory, UriStatsSpanRecorderFactory } = require('../../lib/metric/uri/span-recorder-factory')
+const { SpanRecorderFactory } = require('../../lib/context/trace/span-recorder-factory')
+const { UriStatsSpanRecorderFactory } = require('../../lib/metric/uri/uri-stats-span-recorder-factory')
 const { TraceCompletionEnricher } = require('../../lib/metric/uri/trace-completion-enricher')
 const { ErrorAnalysisConfigBuilder } = require('../../lib/context/trace/error-analysis-config-builder')
 const { ExceptionEnricher, exceptionEnricherNullObject } = require('../../lib/context/trace/exception-enricher')
